Sprint-2 POS application completed.

+++ src/main/java/com/ust/pos/role/service/impl/RoleServiceImpl.java
@@ -7,6 +7,8 @@ import com.ust.pos.role.service.RoleService;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,7 +30,6 @@ public class RoleServiceImpl implements RoleService {
         Role role = roleRepository.findByIdentifier(identifier);
 
         if (role == null) {
-
             return null;
         }
 
@@ -37,42 +38,59 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public RoleDto save(RoleDto roleDto) {
+
         String identifier = roleDto.getIdentifier();
         Role existingRole = roleRepository.findByIdentifier(identifier);
+
         if (existingRole != null) {
             roleDto.setMessage("Role with identifier - " + identifier + " already exists");
             roleDto.setSuccess(false);
             return roleDto;
         }
+
         Role role = modelMapper.map(roleDto, Role.class);
         roleRepository.save(role);
+
         return roleDto;
     }
 
     @Override
     public RoleDto update(RoleDto roleDto) {
+
         String identifier = roleDto.getIdentifier();
         Role existingRole = roleRepository.findByIdentifier(identifier);
+
         if (existingRole == null) {
             roleDto.setMessage("Role with identifier - " + identifier + " not found");
             roleDto.setSuccess(false);
             return roleDto;
         }
+
         modelMapper.map(roleDto, existingRole);
         roleRepository.save(existingRole);
+
         return roleDto;
     }
 
     @Override
     @Transactional
-    public void delete(String identifier) {
+    public boolean delete(String identifier) {
+
+        if (identifier == null) {
+            return false;
+        }
+
         roleRepository.deleteByIdentifier(identifier);
+        return true;
     }
 
     @Override
-    public List<RoleDto> findAll() {
+    public List<RoleDto> findAll(Pageable pageable) {
+
         Type listType = new TypeToken<List<RoleDto>>() {
         }.getType();
-        return modelMapper.map(roleRepository.findAll(), listType);
+        Page<Role> rolePage = roleRepository.findAll(pageable);
+
+        return modelMapper.map(rolePage.getContent(), listType);
     }
 }
